fix SQL VACUUM statements in tests [rt#118460,124854]
Since SQLite 3.15, VACUUM either needs no arguments, or the name of an
attached database.

From https://sqlite.org/lang_vacuum.html:

Compatibility Warning: The ability to vacuum attached databases was
added in version 3.15.0 (2016-10-14). Prior to that, a schema-name
added to the VACUUM statement would be silently ignored and the "main"
schema would be vacuumed.